+++ src/components/Table/components/TableHeaderCell/TableHeaderCell.tsx
@@ -4,6 +4,7 @@ import { flexRender, Header } from '@tanstack/react-table';
 import React from 'react';
 
 import { ACTIONS_COLUMN_ID, TEST_IDS } from '@/constants';
+import { ColumnHeaderFontSize } from '@/types';
 
 import { getStyles } from './TableHeaderCell.styles';
 import { TableHeaderCellFilter } from './TableHeaderCellFilter';
@@ -16,18 +17,25 @@ interface Props<TData> {
    * Header
    */
   header: Header<TData, unknown>;
+
+  /**
+   * Size
+   *
+   * @type {ColumnHeaderFontSize}
+   */
+  size: ColumnHeaderFontSize;
 }
 
 /**
  * Table Header Cell
  */
-export const TableHeaderCell = <TData,>({ header }: Props<TData>) => {
+export const TableHeaderCell = <TData,>({ header, size }: Props<TData>) => {
   /**
    * Styles
    */
   const styles = useStyles2(getStyles);
-
   const sort = header.column.getIsSorted();
+  const fontColor = header.column.columnDef.meta?.config.appearance.header.fontColor || 'inherit';
 
   /**
    * Actions Header
@@ -43,17 +51,21 @@ export const TableHeaderCell = <TData,>({ header }: Props<TData>) => {
         className={cx({
           [styles.labelSortable]: header.column.getCanSort(),
         })}
+        style={{
+          color: fontColor,
+        }}
         {...TEST_IDS.tableHeaderCell.root.apply()}
       >
         {flexRender(header.column.columnDef.header, header.getContext())}
         {!!sort && (
           <Icon
             name={sort === 'asc' ? 'arrow-up' : 'arrow-down'}
+            size={size}
             {...TEST_IDS.tableHeaderCell.sortIcon.apply(sort === 'asc' ? 'arrow-up' : 'arrow-down')}
           />
         )}
       </div>
-      {header.column.columnDef.enableColumnFilter && <TableHeaderCellFilter header={header} />}
+      {header.column.columnDef.enableColumnFilter && <TableHeaderCellFilter header={header} size={size} />}
     </>
   );
 };